Tests/LibWeb/Text/expected/input-file-accept.txt

Select file...file1 Select files...4 files selected. Select file...file1.cpp Select files...2 files selected.   input1:
file1: text/plain: Contents for file1
input2:
file1: text/plain: Contents for file1
file2: text/plain: Contents for file2
file3: text/plain: Contents for file3
file4: text/plain: Contents for file4
input3:
file1.cpp: text/plain: int main() {{ return 1; }}
input4:
file1.cpp: text/plain: int main() {{ return 1; }}
file2.cpp: text/plain: int main() {{ return 2; }}
